[POST] Exception 처리

+++ src/main/java/com/swjungle/board/post/service/PostService.java
@@ -6,9 +6,13 @@ import com.swjungle.board.post.dto.request.CreatePostRequest;
 import com.swjungle.board.post.dto.response.PostResponse;
 import com.swjungle.board.post.dto.request.UpdatePostRequest;
 import com.swjungle.board.post.entity.Post;
+import com.swjungle.board.post.exception.InvalidPostRequestException;
+import com.swjungle.board.post.exception.PostNotFoundException;
 import com.swjungle.board.post.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.BeanPropertyBindingResult;
+import org.springframework.validation.Errors;
 
 import java.util.List;
 
@@ -16,9 +20,20 @@ import java.util.List;
 @RequiredArgsConstructor
 public class PostService {
     private final PostRepository postRepository;
+    private final PostRequestValidator postRequestValidator;
 
     public PostResponse createPost(CreatePostRequest createPostRequest) {
-        Post post = Post.of(createPostRequest.title(), createPostRequest.content(), createPostRequest.link(), createPostRequest.category(), createPostRequest.score(), createPostRequest.author(),createPostRequest.password());
+        postRequestValidator.validate(createPostRequest); // 유효성 검증
+
+        Post post = Post.of(
+                createPostRequest.title(),
+                createPostRequest.content(),
+                createPostRequest.link(),
+                createPostRequest.category(),
+                createPostRequest.score(),
+                createPostRequest.author(),
+                createPostRequest.password()
+        );
         Post newPost = postRepository.save(post);
         return PostResponse.fromEntity(newPost);
     }
@@ -28,17 +43,38 @@ public class PostService {
     }
 
     public PostResponse getPostById(Long id) {
-        return PostResponse.fromEntity(postRepository.findById(id).orElse(null));
+        return PostResponse.fromEntity(postRepository.findById(id).orElseThrow(() -> new PostNotFoundException(id)));
     }
 
     public PostResponse updatePost(Long id, UpdatePostRequest updatePostRequest) {
         Post existingPost = postRepository.getReferenceById(id);
-        existingPost.update(updatePostRequest.title(), updatePostRequest.content(), updatePostRequest.link(), updatePostRequest.category(), updatePostRequest.score(), updatePostRequest.author(), updatePostRequest.password());
+
+        if (existingPost == null) {
+            throw new PostNotFoundException(id);
+        }
+
+        postRequestValidator.validate(updatePostRequest); // 유효성 검증
+
+        existingPost.update(
+                updatePostRequest.title(),
+                updatePostRequest.content(),
+                updatePostRequest.link(),
+                updatePostRequest.category(),
+                updatePostRequest.score(),
+                updatePostRequest.author(),
+                updatePostRequest.password()
+        );
+
         return PostResponse.fromEntity(postRepository.save(existingPost));
     }
 
     public MessageResponse deletePost(Long id) {
         Post existingPost = postRepository.getReferenceById(id);
+
+        if (existingPost == null) {
+            throw new PostNotFoundException(id);
+        }
+
         postRepository.deleteById(id);
         return new MessageResponse("삭제 완료");
     }
